refactored linter and configuration-manager
