+++ src/extension.ts
@@ -20,59 +20,23 @@ import { EllipsesFormattingProvider } from "./typeFormatters/ellipsesFormatter";
 import { OnTypeFormattingDispatcher } from './typeFormatters/dispatcher';
 import { QuotesFormattingProvider } from './typeFormatters/quotesFormatter';
 import { LT_DOCUMENT_SELECTORS, LT_OUTPUT_CHANNEL, LT_TIMEOUT_MS, LT_SERVICE_MANAGED } from './common/constants';
-import { ConfigurationManager } from "./common/configuration";
-import { LinterCommands } from "./linter/commands";
-
-const config: ConfigurationManager = new ConfigurationManager();
-const linter: LinterCommands = new LinterCommands(config);
-const onTypeDispatcher = new OnTypeFormattingDispatcher({
-  '"': new QuotesFormattingProvider(config),
-  "'": new QuotesFormattingProvider(config),
-  '-': new DashesFormattingProvider(config),
-  '.': new EllipsesFormattingProvider(config)
-});
-const onTypeTriggers = onTypeDispatcher.getTriggerCharacters();
-
-// CodeActionProvider
-class LTCodeActionProvider implements vscode.CodeActionProvider {
-  private readonly linter: LinterCommands;
-
-  constructor(linter: LinterCommands) {
-    this.linter = linter;
-  }
-
-  provideCodeActions(
-    document: vscode.TextDocument,
-    range: vscode.Range,
-    context: vscode.CodeActionContext,
-    token: vscode.CancellationToken
-  ): vscode.CodeAction[] {
-    let documentUri: string = document.uri.toString();
-    // let diagnosticMap: Map<string, vscode.Diagnostic[]> = this.linter.getDiagnosticMap();
-    let codeActionMap: Map<string, vscode.CodeAction[]> = this.linter.getCodeActionMap();
-    if (codeActionMap.has(documentUri) && codeActionMap.get(documentUri)) {
-      let documentCodeActions: vscode.CodeAction[] = codeActionMap.get(documentUri) || [];
-      let actions: vscode.CodeAction[] = [];
-      // Code Actions get created in suggest()
-      documentCodeActions.forEach(function (action) {
-        if (action.diagnostics && context.diagnostics) {
-          let actionDiagnostic: vscode.Diagnostic = action.diagnostics[0];
-          if (range.contains(actionDiagnostic.range)) {
-            actions.push(action);
-          }
-        }
-      });
-      return actions;
-    } else {
-      return [];
-    }
-  }
-}
+import { ConfigurationManager } from "./common/configuration-manager";
+import { Linter } from "./linter/linter";
 
 // Wonder Twin Powers, Activate!
 export function activate(context: vscode.ExtensionContext) {
 
-  context.subscriptions.push(config);
+  const configMan: ConfigurationManager = new ConfigurationManager();
+  const linter: Linter = new Linter(configMan);
+  const onTypeDispatcher = new OnTypeFormattingDispatcher({
+    '"': new QuotesFormattingProvider(configMan),
+    "'": new QuotesFormattingProvider(configMan),
+    '-': new DashesFormattingProvider(configMan),
+    '.': new EllipsesFormattingProvider(configMan)
+  });
+  const onTypeTriggers = onTypeDispatcher.getTriggerCharacters();
+
+  context.subscriptions.push(configMan);
 
   context.subscriptions.push(LT_OUTPUT_CHANNEL);
   LT_OUTPUT_CHANNEL.appendLine("LanguageTool Linter Activated!");
@@ -80,7 +44,7 @@ export function activate(context: vscode.ExtensionContext) {
   // Register onDidChangeconfiguration event
   context.subscriptions.push(vscode.workspace.onDidChangeConfiguration(event => {
     if (event.affectsConfiguration("languageToolLinter")) {
-      config.reloadConfiguration(event);
+      configMan.reloadConfiguration(event);
     }
   }));
 
@@ -104,7 +68,7 @@ export function activate(context: vscode.ExtensionContext) {
   // Register onDidChangeTextDocument event - request lint with default timeout
   context.subscriptions.push(vscode.workspace.onDidChangeTextDocument(event => {
     // if (ltConfig.get("lintOnChange")) {
-    if (config.getLintOnChange()) {
+    if (configMan.getLintOnChange()) {
       linter.requestLint(event.document);
       // requestLint(event.document);
     }
@@ -119,7 +83,7 @@ export function activate(context: vscode.ExtensionContext) {
   // Register Code Actions Provider for supported languages
   LT_DOCUMENT_SELECTORS.forEach(function (selector: vscode.DocumentSelector) {
     context.subscriptions.push(
-      vscode.languages.registerCodeActionsProvider(selector, new LTCodeActionProvider(linter))
+      vscode.languages.registerCodeActionsProvider(selector, linter)
     );
 
     if (onTypeTriggers) {
@@ -148,7 +112,7 @@ export function activate(context: vscode.ExtensionContext) {
 
   // Register "Auto Format Document" TextEditorCommand
   let autoFormatCommand = vscode.commands.registerTextEditorCommand("languagetoolLinter.autoFormatDocument", (editor, edit) => {
-    if (config.isSupportedDocument(editor.document)) {
+    if (configMan.isSupportedDocument(editor.document)) {
       // Revert to regex here for cleaner code.
       let text: string = editor.document.getText();
       let lastOffset: number = text.length - 1;
@@ -171,7 +135,7 @@ export function activate(context: vscode.ExtensionContext) {
   // Lint Active Text Editor on Activate
   if (vscode.window.activeTextEditor) {
     let firstDelay = LT_TIMEOUT_MS;
-    if (config.getServiceType() === LT_SERVICE_MANAGED) {
+    if (configMan.getServiceType() === LT_SERVICE_MANAGED) {
       // Add a second to give the service time to start up.
       firstDelay += 1000;
     }
