+++ src/test/suite/configuration.test.ts
@@ -1,7 +1,7 @@
 import * as assert from 'assert';
 import { before } from 'mocha';
 import * as vscode from 'vscode';
-import { ConfigurationManager } from '../../common/configuration';
+import { ConfigurationManager } from '../../common/configuration-manager';
 import * as constants from '../../common/constants';
 
 suite('ConfigurationManager Test Suite', () => {
