+++ src/test/suite/extension.test.ts
@@ -1,8 +1,8 @@
 import * as assert from 'assert';
 import { before } from 'mocha';
 import * as vscode from 'vscode';
-import * as linter from '../../linter/commands';
-import * as configman from '../../common/configuration';
+import * as linter from '../../linter/linter';
+import * as configman from '../../common/configuration-manager';
 import * as constants from '../../common/constants';
 
 // import * as languagetooLinter from '../';
